RecordUpdater: code cleanup
